Use React component to add 3d box in the scene
