+++ ios/components/ARBoxManager.h
@@ -1,0 +1,15 @@
+//
+//  ARBoxManager.h
+//  RCTARKit
+//
+//  Created by Zehao Li on 8/12/17.
+//  Copyright © 2017 HippoAR. All rights reserved.
+//
+
+#import <React/RCTBridgeModule.h>
+
+@interface ARBoxManager : NSObject <RCTBridgeModule>
+
+@end
+
+
